Pull everything up into IDag
